separating unique uli function
